third_party/blink/web_tests/wpt_internal/css/css-conditional/container-queries/html-root-container-writing-mode.html

<!doctype html>
<title>The body writing-mode propagated to html root for container queries</title>
<link rel="help" href="https://drafts.csswg.org/css-writing-modes/#principal-flow">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
  html { contain: size layout style; }
  body { writing-mode: vertical-lr; margin: 0; }
  @container (min-height: 50vh) {
    div { color: green }
  }
  @container (min-width: 50vw) {
    div { color: red }
  }
</style>
<div id="green"></div>
<script>
  test(() => {
    assert_equals(document.documentElement.offsetLeft, 0, "offsetLeft");
    assert_equals(document.documentElement.offsetWidth, 0, "offsetWidth");
    assert_equals(document.documentElement.offsetHeight, window.innerHeight, "offsetHeight");
  }, "The html root element is laid out in vertical-lr");

  test(() => {
    assert_equals(document.body.offsetLeft, 0, "offsetLeft");
    assert_equals(document.body.offsetWidth, 0, "offsetWidth");
    assert_equals(document.body.offsetHeight, window.innerHeight, "offsetHeight");
  }, "The body element is laid out in vertical-lr");

  test(() => {
    assert_equals(getComputedStyle(green).color, "rgb(0, 128, 0)", "div in container should be green");
  }, "The div element responds to html container in correct orientation");
</script>
